Improve dynamic partition flashing support -Now dynamic partitions can be resized before flashing if necessary -Requires /system_image or /vendor_image mount point in twrp.flags -Add new feature: "Free up space in Super" - removes all non-main dynamic partitions (system_ext, product, my_*) -Add new feature: "Convert image to r/w" - unshares blocks and resizes image before flashing -Use phhusson's lptools to manage dynamic partitions

+++ twrp-functions.cpp
@@ -1876,7 +1876,7 @@ void TWFunc::SetPerformanceMode(bool mode)
   usleep(500000);
 }
 
-std::string TWFunc::to_string(unsigned long value)
+std::string TWFunc::to_string(unsigned long long value)
 {
   std::ostringstream os;
   os << value;
